adjustments on the log viewer, removed the live viewer added a reload button, brought back the log level selector.
