+++ web/src/screens/settings/Logs.tsx
@@ -1,14 +1,15 @@
 import {useMutation, useQuery, useQueryClient} from "@tanstack/react-query";
 import {ConfigQueryOptions, LogQueryOptions} from "@api/queries.ts";
-import {CenteredEmptyState, SettingsSectionHeader} from "@screens/settings/components.tsx";
-import { Stack, Text, Select } from "@mantine/core";
+import {CenteredEmptyState, SettingsSectionHeader, LogViewer} from "@screens/settings/components.tsx";
+import { Divider, Stack, Text, Select, Group, ActionIcon } from "@mantine/core";
 import { useState } from "react";
 import {APIClient} from "@api/APIClient.ts";
 import {displayNotification} from "@components/notifications";
 import {SettingsKeys} from "@api/query_keys.ts";
+import {FaDownload} from "react-icons/fa";
+import {baseUrl} from "@utils";
 
 export const Logs = () => {
-
     const queryClient = useQueryClient();
     const {data} = useQuery(ConfigQueryOptions());
 
@@ -28,47 +29,103 @@ export const Logs = () => {
         }
     });
 
-    // Always set log level to TRACE on mount
-    useState(() => {
-        if (data && data.log_level !== "TRACE") {
-            mutation.mutate({ log_level: "TRACE" });
-        }
-    });
     return (
         <main>
             <SettingsSectionHeader title={"Logs"} description={"Manage shinkro logs here."}/>
+            
+            {/* Log Settings Section */}
+            {data && (
+                <Stack mt={"md"}>
+                    <Text fw={900} size={"xl"}>
+                        Log Settings:
+                    </Text>
+                    <Group>
+                        <Text fw={600} size={"md"}>
+                            Log Level:
+                        </Text>
+                        <Select
+                            size={"xs"}
+                            data={[
+                                {value: "DEBUG", label: "DEBUG"},
+                                {value: "INFO", label: "INFO"},
+                                {value: "ERROR", label: "ERROR"},
+                                {value: "TRACE", label: "TRACE"},
+                            ]}
+                            defaultValue={data.log_level}
+                            onChange={(value: LogLevel | string | null) => mutation.mutate({log_level: value ? value as LogLevel : ""})}
+                        />
+                    </Group>
+                </Stack>
+            )}
+            
+            <Divider mt={"md"}/>
+            
+            {/* Log Viewer Section */}
+            <Stack mt={"md"}>
+                <Text fw={900} size={"xl"}>
+                    Log Viewer:
+                </Text>
+                <LogViewer />
+            </Stack>
+            
+            <Divider mt={"md"}/>
+            
+            {/* Log Files Section */}
             <LogFiles/>
         </main>
     );
 }
 
-// ...existing code...
-import { LogViewer } from "./LogViewer";
-
 export const LogFiles = () => {
-    const { data: logs } = useQuery(LogQueryOptions());
-    const [selected, setSelected] = useState<string | null>(null);
-    const isEmpty = !logs || !(logs.length > 0);
+    const {data: logs} = useQuery(LogQueryOptions());
+    const [isDownloading, setIsDownloading] = useState(false);
+    const isEmpty = !logs || !(logs.length > 0)
+
+    const handleDownload = async (filename: string) => {
+        setIsDownloading(true);
+        const response = await fetch(`${baseUrl()}api/fs/logs/${filename}`);
+        const blob = await response.blob();
+        const url = URL.createObjectURL(blob);
+        const link = document.createElement("a");
+        link.href = url;
+        link.download = filename;
+        link.click();
+        URL.revokeObjectURL(url);
+        setIsDownloading(false);
+    }
 
     return (
         <main>
             <Text fw={900} size={"xl"} mt={"md"}>
-                Log Viewer:
+                Log Files:
             </Text>
             {isEmpty ? (
-                <CenteredEmptyState message={"No Logs Found"} />
+                <CenteredEmptyState message={"No Logs Found"}/>
             ) : (
                 <Stack>
-                    <Select
-                        label="Log file"
-                        data={logs.map((log) => ({ value: log.name, label: log.name }))}
-                        value={selected}
-                        onChange={setSelected}
-                        placeholder="Select a log file"
-                        size="xs"
-                        mb="sm"
-                    />
-                    {selected && <LogViewer filename={selected} />}
+                    {logs.map((log) => (
+                        <Group key={log.name} align={"flex-start"} mt={"md"}>
+                            <Stack gap={0}>
+                                <Text fw={600}>
+                                    {log.name}
+                                </Text>
+                                <Text size={"xs"} c={"dimmed"}>
+                                    Size: {log.size_human}
+                                </Text>
+                                <Text size={"xs"} c={"dimmed"}>
+                                    Last Modified: {new Date(log.modified_at).toLocaleString()}
+                                </Text>
+                            </Stack>
+                            <Stack>
+                                <ActionIcon
+                                    loading={isDownloading}
+                                    onClick={() => handleDownload(log.name)}
+                                >
+                                    <FaDownload/>
+                                </ActionIcon>
+                            </Stack>
+                        </Group>
+                    ))}
                 </Stack>
             )}
         </main>
